config: accept profile name in config patch
This is for setting profile names of the internal $rsprofile profile.
And is also used for initial sync of normal profiles. Previously, the
internal profiles were not named.

Also always create the profile even if sets is empty to avoid a desync
with the client. Previously the client would detect the profile missing
and unmark it as sync if it was created and no keys added to it.